requirements.txt
streamlit
streamlit-js_eval
langchain
langgraph
langchain-openai
langchain-google-genai
python-dotenv
requests
pydantic
plotly
pandas
pysqlite3-binary
numpy<2.0.0
